Analysis README.md
It is such an awesome project by which anyone can realize the power of data analytics in sports. Using 
basic data manipulation, visualizing the results, and deducing the patterns involved, this project will 
allow users to gain actionable insights for the performance of a team, which happens with the ability to 
categorize results, calculate key metrics, and visualize performance trends.